add QQ for nockma terms
